Add settings rollback, created callback service for input changes
Rollback:
after any input is changed users can rollback to default (for particular setting)

Callback Service:
There was a problem with changing inputs, users would have to change them twice for a change to be saved. Callback functions ensure that input is saved again only when it was changed. used callback for all inputs
Created dedicated service file for that

Closes #14